EBMApredict {EBMAforecast}	R Documentation

EBMApredict

Description

Function allows users to create new predictions given an already estimated EBMA model This function produces predictions based on EBMA model weights and component model predictions.

Usage

EBMApredict(EBMAmodel, Predictions, Outcome = NULL, ...)

## S4 method for signature 'ForecastData'
EBMApredict(EBMAmodel, Predictions, Outcome = NULL,
  ...)

prediction(EBMAmodel, Predictions, Outcome, ...)

## S4 method for signature 'FDatFitLogit'
prediction(EBMAmodel, Predictions, Outcome, ...)

## S4 method for signature 'FDatFitNormal'
prediction(EBMAmodel, Predictions, Outcome = c(),
  ...)

Arguments

EBMAmodel
Output of estimated EBMA model

Predictions
A matrix with a column for each component model's predictions.

Outcome
An optional vector containing the true values of the dependent variable for all observations in the test period.

...
Not implemented

Value

Returns a data of class 'FDatFitLogit' or FDatFitNormal, a subclass of 'ForecastData', with the following slots:

predTest
A matrix containing the predictions of all component models and the EBMA model for all observations in the test period.

#'

period
The period, "calibration" or "test", for which the statistics were calculated.

outcomeTest
An optional vector containing the true values of the dependent variable for all observations in the test period.

modelNames
A character vector containing the names of all component models. If no model names are specified, names will be assigned automatically.

modelWeights
A vector containing the model weights assigned to each model.

Author(s)

Michael D. Ward <michael.d.ward@duke.edu> and Jacob M. Montgomery <jacob.montgomery@wustl.edu> and Florian M. Hollenbach <florian.hollenbach@tamu.edu>

References

Montgomery, Jacob M., Florian M. Hollenbach and Michael D. Ward. (2015). Calibrating ensemble forecasting models with sparse data in the social sciences. International Journal of Forecasting. In Press.

Montgomery, Jacob M., Florian M. Hollenbach and Michael D. Ward. (2012). Improving Predictions Using Ensemble Bayesian Model Averaging. Political Analysis. 20: 271-291.

[Package EBMAforecast version 0.52 Index]